move hittable and related to hittable.hpp
